+++ Lib9c/Model/State/ShopStateAlreadyContainsException.cs
@@ -1,0 +1,19 @@
+using System;
+using System.Runtime.Serialization;
+
+namespace Nekoyume.Model.State
+{
+    [Serializable]
+    public class ShopStateAlreadyContainsException : Exception
+    {
+        public ShopStateAlreadyContainsException(string message) : base(message)
+        {
+        }
+
+        protected ShopStateAlreadyContainsException(
+            SerializationInfo info,
+            StreamingContext context) : base(info, context)
+        {
+        }
+    }
+}
